Define MemoryEffectOpInterface for all torch ops (#9)
Defines a MemoryEffectOpInterface default implementation for all torch ops. The implementation ensures that torch ops that only operate on tensors with Value semantics are marked as having no side effects, and operations that operate on tensors with non-value semantics are marked as conservatively having both read and write side effects. This enables folding value-semantics ops that are unused.

Signed-off-by: James Bartlett <james@gimletlabs.ai>